modularized for code challenge and readability
